pyproject.toml
[project]
name = "autodev-repair-bot"
version = "0.1.0"
description = "Agentic CI/CD repair bot with sandboxed verification"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Eddy", email = "your-email@example.com" }
]
keywords = ["agentic-ai", "ci-cd", "devtools", "llm", "automation", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "langgraph>=0.0.45",
    "langchain-openai>=0.1.7",
    "docker>=7.0.0",
    "gitpython>=3.1.40",
    "pydantic>=2.6.0",
    "rich>=13.7.0",
    "unidiff>=0.7.5",
    "PyGithub>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
autodev = "autodev.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/autodev-repair-bot"
Documentation = "https://github.com/yourusername/autodev-repair-bot#readme"
Repository = "https://github.com/yourusername/autodev-repair-bot.git"
Issues = "https://github.com/yourusername/autodev-repair-bot/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["autodev*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
